feat: add ggircs backup dag & factor all dags into one file with a function call
